move GAME_ROUND to its own file
