+++ include/game/point.h
@@ -1,0 +1,18 @@
+#ifndef POINT_H
+#define POINT_H
+
+static const BG_POINT CUR_BLIND_TOKEN_POS = {8, 18};
+
+// Game points
+static const BG_POINT HELD_JOKERS_POS = {108, 10};
+static const BG_POINT JOKER_DISCARD_TARGET = {240, 30};
+
+// Round phase points
+static const BG_POINT CARD_DRAW_POS = {208, 110};
+static const BG_POINT CARD_DISCARD_PNT = {240, 70};
+static const BG_POINT HAND_START_POS = {120, 90};
+static const BG_POINT HAND_PLAY_POS = {120, 70};
+static const BG_POINT PLAYING_SCORE_FLAME_CHIPS_POS __attribute__((unused)) = {176, 80};
+static const BG_POINT PLAYING_SCORE_FLAME_MULT_POS __attribute__((unused)) = {208, 80};
+
+#endif // POINT_H
