+++ include/game/round.h
@@ -1,0 +1,10 @@
+#ifndef __INCLUDE_GAME_ROUND_H__
+#define __INCLUDE_GAME_ROUND_H__
+
+#include <tonc.h>
+
+// Main round state functions
+void game_round_on_init(void);
+void game_playing_on_update(void);
+
+#endif
