+++ include/game/selection.h
@@ -1,0 +1,9 @@
+#ifndef SELECTION_H
+#define SELECTION_H
+
+#include "selection_grid.h"
+
+// Selection grid positions
+static const Selection GAME_PLAYING_INIT_SEL = {0, 1};
+
+#endif // SELECTION_H
